requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
black==21.12b0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
Django==4.0.1
django-js-asset==1.2.2
django-mptt==0.13.4
factory-boy==3.2.1
Faker==11.3.0
flake8==4.0.1
h11==0.12.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
outcome==1.1.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pytest==6.2.5
pytest-base-url==1.4.2
pytest-django==4.5.2
pytest-factoryboy==2.1.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-selenium==2.0.1
pytest-variables==1.9.0
python-dateutil==2.8.2
requests==2.27.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tenacity==6.3.1
text-unidecode==1.3
toml==0.10.2
tomli==1.2.3
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.8
webdriver-manager==3.5.2
wsproto==1.0.0
